talis: Don't collect RIS files
The RIS and HTML sources are similar enough that they could be hard to
de-duplicate.  In addition, while we could theoretically benefit from
the nicely structured format of the RIS files, we don't currently have
any system in place to benefit from that structure, and mixing
structured and unstructured data adds complexity.  For now, stick with
just HTML.