Simplify requirements.txt
There's no need to have requirements.txt explicitly detail every
dependency -- we can let `pip` take care of most of that.  Additionally,
remove strict version specifications (`==`) for everything but Scrapy,
since most of the time we benefit from having the most recent packages.
The packages `ipython` and `rq-dashboard` aren't strictly necessary, but
are helpful.

Made with the help of `pigar`.