Add filter to set absolute max crawl depth
Adding this filter allows us to avoid endlessly crawling sites that
have an infinite number of procedurally generated pages with unique
URLs.  The correct depth to use is hard to determine, as some sites
with paginated search results have valuable content at a very high
depth, so this is left as a parameter that can be specified with a
parameter when starting a crawl.