Run crawls for each row in sequence
In our crawls, we try to minimize the impact of a crawl on each site,
but we run multiple crawls against the same site at the same time.
This change causes the crawl jobs in each row to run in sequence
instead of (potentially) in parallel.  Each job will wait for the
previous job to finish before starting.