Prevent invalid scraper name causing job failure
Currently, an invalid spider name in the Custom Scraper Name column
will cause the job to fail, which prevents subsequent jobs that
depend on it from executing.  With this change, invalid scraper
names will cause the job to log a warning that the scraper name
is invalid and the job will finish successfully.