Update settings for broad crawls
The 'DOWNLOAD_TIMEOUT' setting has been added as part of a
`custom_settings` definition in FilterSpider, rather than in
`settings/scrapy.py`, because the Scrapy default might still be
appropriate for other types of spiders, i.e., children of CustomSpider.

This does prompt the question of where settings should be stored.
Should we lean on the global settings file, or on definitions within the
classes?  For now, I think what we have makes sense, but it's worth
revisiting in the future.